+++ sdk/textanalytics/azure-ai-textanalytics/src/test/java/com/azure/ai/textanalytics/TestUtils.java
@@ -10,6 +10,9 @@ import com.azure.ai.textanalytics.implementation.AnalyzeHealthcareEntitiesResult
 import com.azure.ai.textanalytics.implementation.AnalyzeSentimentActionResultPropertiesHelper;
 import com.azure.ai.textanalytics.implementation.AssessmentSentimentPropertiesHelper;
 import com.azure.ai.textanalytics.implementation.CategorizedEntityPropertiesHelper;
+import com.azure.ai.textanalytics.implementation.ClassificationCategoryPropertiesHelper;
+import com.azure.ai.textanalytics.implementation.ClassifyDocumentResultPropertiesHelper;
+import com.azure.ai.textanalytics.implementation.DynamicClassifyDocumentResultCollectionPropertiesHelper;
 import com.azure.ai.textanalytics.implementation.ExtractKeyPhrasesActionResultPropertiesHelper;
 import com.azure.ai.textanalytics.implementation.ExtractSummaryActionResultPropertiesHelper;
 import com.azure.ai.textanalytics.implementation.ExtractSummaryResultPropertiesHelper;
@@ -35,6 +38,8 @@ import com.azure.ai.textanalytics.models.AnalyzeSentimentResult;
 import com.azure.ai.textanalytics.models.AssessmentSentiment;
 import com.azure.ai.textanalytics.models.CategorizedEntity;
 import com.azure.ai.textanalytics.models.CategorizedEntityCollection;
+import com.azure.ai.textanalytics.models.ClassificationCategory;
+import com.azure.ai.textanalytics.models.ClassifyDocumentResult;
 import com.azure.ai.textanalytics.models.DetectLanguageInput;
 import com.azure.ai.textanalytics.models.DetectLanguageResult;
 import com.azure.ai.textanalytics.models.DetectedLanguage;
@@ -77,6 +82,7 @@ import com.azure.ai.textanalytics.models.TextSentiment;
 import com.azure.ai.textanalytics.util.AnalyzeHealthcareEntitiesResultCollection;
 import com.azure.ai.textanalytics.util.AnalyzeSentimentResultCollection;
 import com.azure.ai.textanalytics.util.DetectLanguageResultCollection;
+import com.azure.ai.textanalytics.util.DynamicClassifyDocumentResultCollection;
 import com.azure.ai.textanalytics.util.ExtractKeyPhrasesResultCollection;
 import com.azure.ai.textanalytics.util.ExtractSummaryResultCollection;
 import com.azure.ai.textanalytics.util.RecognizeEntitiesResultCollection;
@@ -129,6 +135,10 @@ final class TestUtils {
     static final List<String> CUSTOM_MULTI_CLASSIFICATION = asList(
         "I need a reservation for an indoor restaurant in China. Please don't stop the music. Play music and add"
             + " it to my playlist");
+
+    static final List<String> DYNAMIC_CLASSIFICATION = asList(
+        "The WHO is issuing a warning about Monkey Pox.",
+        "Mo Salah plays in Liverpool FC in England.");
 
     static final List<String> SUMMARY_INPUTS = asList(
         "At Microsoft, we have been on a quest to advance AI beyond existing techniques, by taking a more holistic,"
@@ -1386,6 +1396,43 @@ final class TestUtils {
         SummarySentencePropertiesHelper.setOffset(summarySentence, offset);
         SummarySentencePropertiesHelper.setLength(summarySentence, length);
         return summarySentence;
+    }
+
+    // Dynamic classification
+    static DynamicClassifyDocumentResultCollection getExpectedDynamicClassifyDocumentResultCollection() {
+        ClassifyDocumentResult classifyDocumentResult1 = new ClassifyDocumentResult("0",
+            new TextDocumentStatistics(46, 1), null);
+
+        ClassificationCategory classificationCategory1 = new ClassificationCategory();
+        ClassificationCategoryPropertiesHelper.setCategory(classificationCategory1, "Health");
+        ClassificationCategoryPropertiesHelper.setConfidenceScore(classificationCategory1, 0.88);
+        ClassificationCategory classificationCategory2 = new ClassificationCategory();
+        ClassificationCategoryPropertiesHelper.setCategory(classificationCategory2, "Music");
+        ClassificationCategoryPropertiesHelper.setConfidenceScore(classificationCategory2, 0.04);
+
+        ClassificationCategory classificationCategory3 = new ClassificationCategory();
+        ClassificationCategoryPropertiesHelper.setCategory(classificationCategory3, "Sports");
+        ClassificationCategoryPropertiesHelper.setConfidenceScore(classificationCategory3, 0.04);
+
+        ClassificationCategory classificationCategory4 = new ClassificationCategory();
+        ClassificationCategoryPropertiesHelper.setCategory(classificationCategory4, "Politics");
+        ClassificationCategoryPropertiesHelper.setConfidenceScore(classificationCategory4, 0.03);
+
+        ClassifyDocumentResultPropertiesHelper.setClassifications(classifyDocumentResult1,
+            IterableStream.of(asList(classificationCategory1, classificationCategory2, classificationCategory3,
+            classificationCategory4)));
+
+        ClassifyDocumentResult classifyDocumentResult2 = new ClassifyDocumentResult("1",
+            new TextDocumentStatistics(42, 1), null);
+        ClassifyDocumentResultPropertiesHelper.setClassifications(classifyDocumentResult2,
+            IterableStream.of(asList(classificationCategory3, classificationCategory2, classificationCategory1,
+                classificationCategory4)));
+
+        DynamicClassifyDocumentResultCollection classifyDocumentResults =
+            new DynamicClassifyDocumentResultCollection(asList(classifyDocumentResult1, classifyDocumentResult2));
+        DynamicClassifyDocumentResultCollectionPropertiesHelper.setStatistics(classifyDocumentResults,
+            new TextDocumentBatchStatistics(2, 2, 0, 0));
+        return classifyDocumentResults;
     }
 
     /**